fix ui issue and add translation
