+++ public/locales/en-US/page-downloads.json
@@ -23,6 +23,7 @@
   },
   "DownloadSection": {
     "title": "Downloads",
+    "subtitle": "The media assets are released under CCO license. \"No Rights Reserved\".",
     "entries": [
       {
         "name": "DeFiChain logo",
