+++ public/locales/fr/page-downloads.json
@@ -23,6 +23,7 @@
   },
   "DownloadSection": {
     "title": "Téléchargements",
+    "subtitle": "The media assets are released under CCO license. \"No Rights Reserved\".",
     "entries": [
       {
         "name": "Logo de DeFiChain",
