+++ public/locales/zh-Hans/page-downloads.json
@@ -23,6 +23,7 @@
   },
   "DownloadSection": {
     "title": "下載",
+    "subtitle": "The media assets are released under CCO license. \"No Rights Reserved\".",
     "entries": [
       {
         "name": "递飞链标志",
